[FEATURE] PlayView
Se arreglo bug de seleccionar poblados(ahora no se puede selecionar poblados que tienen dueño). Se implemento label para puntaje de jugador, se implemento labels e imagenes para representar la cantidad de cartas de desarrollo del jugador, ya se pueden comprar ciudades y se imprimen en pantalla

+++ Catan/PlayView.cpp
@@ -34,7 +34,12 @@ void PlayView::createLabels() {
   cards = new Label("Componentes", sf::Color(0, 0, 255, 128), font,
                     sf::Text::Bold, 20, 565.f, 575.f);
 }
-
+void PlayView::createLabelScorePlayer() {
+    titleScorePlayer = new Label("Score:", sf::Color(0, 0, 255, 128), font,
+        sf::Text::Bold, 20, 1000.f, 30.f);
+    scorePlayer = new Label(std::to_string((*game.playerIterator)->getScore()), sf::Color(0, 0, 255, 128), font,
+        sf::Text::Bold, 20, 1100.f, 30.f);
+}
 void PlayView::createLabelNamePlayers() {
   playerIterator = beginPlayerIterator();
   if (isPlayerListTraversal()) {
@@ -214,7 +219,24 @@ void PlayView::createLabelFigurePlayer() {
         sf::Color(0, 0, 255, 128), font, sf::Text::Bold, 20,
         680.f, 620.f);
 }
+void PlayView::printPlayerDevelopCard() {
+    printImages("Images/playerCard/mini_progressCard1.png", 720, 640);
+    printImages("Images/playerCard/mini_knightCard1.png", 760, 640);
+    printImages("Images/playerCard/mini_victoryPointCard1.png", 800, 640);
+}
+void PlayView::createLabelPlayerDevelopCard() {
+    victory = new Label(std::to_string((*game.playerIterator)->victoryPointsCards->size()),
+        sf::Color(0, 0, 255, 128), font, sf::Text::Bold, 20,
+        720.f, 620.f);
 
+    knight = new Label(std::to_string((*game.playerIterator)->knightCards->size()),
+        sf::Color(0, 0, 255, 128), font, sf::Text::Bold, 20,
+        760, 620.f);
+
+    progress = new Label(std::to_string((*game.playerIterator)->progressCards->size()),
+        sf::Color(0, 0, 255, 128), font, sf::Text::Bold, 20,
+        800.f, 620.f);
+}
 /*methods of creating Buttons*/
 
 
@@ -295,11 +317,14 @@ void PlayView::drawLabelNumTurn() {
 
 void PlayView::drawLabels() {
   createLabels();
+  createLabelScorePlayer();
   view.draw(materialCard->getTextInstance());
   view.draw(developCard->getTextInstance());
   view.draw(pricingTable->getTextInstance());
   view.draw(turns->getTextInstance());
   view.draw(cards->getTextInstance());
+  view.draw(titleScorePlayer->getTextInstance());
+  view.draw(scorePlayer->getTextInstance());
 }
 void PlayView::drawLabelNamePlayers() {
   createLabelNamePlayers();
@@ -331,7 +356,12 @@ void PlayView::drawLabelFigurePlayer() {
   view.draw(cityPlayer->getTextInstance());
   view.draw(roadPlayer->getTextInstance());
 }
-
+void PlayView::drawLabelDevelopCardPLayer() {
+    createLabelPlayerDevelopCard();
+    view.draw(knight->getTextInstance());
+    view.draw(progress->getTextInstance());
+    view.draw(victory->getTextInstance());
+}
 /*methods of print Card*/
 void PlayView::printPlayerCard() {
   printImages("Images/playerCard/clayCard.png", 360, 640);
@@ -340,6 +370,7 @@ void PlayView::printPlayerCard() {
   printImages("Images/playerCard/woodCard.png", 480, 640);
   printImages("Images/playerCard/woolCard.png", 520, 640);
 }
+
 
 void PlayView::printPlayerFigure() {
   cityIterator = (*game.playerIterator)->citys->begin();
@@ -370,26 +401,31 @@ void PlayView::traverseLands(double x, double y) {
   }
 }
 
-void PlayView::printTownPlayer(list<Vertex *>::iterator vIterator, int x,
-                               int y) {
-  if ((*game.playerIterator)->towns->size() > 0) {
-    if ((*game.playerIterator)->getTownFirstTurn() < 2) {
-      initializeIteratorTownList();
-      setIsClickedToVertexGraph(vIterator);
-      if (townIterator != (*game.playerIterator)->towns->end())
-        printImages((*townIterator)->getImagePath(),
-                    (*vIterator)->getTown()->getPosX(),
-                    (*vIterator)->getTown()->getPosY());
-      setOwnerToVertexGraph(game.graph.getVertex((*vIterator)->getVertexId()));
-      deleteTowntoPlayer();
-      (*game.playerIterator)->setTownFirstTurn(1);
-      if ((*game.playerIterator)->getTownFirstTurn() == 2) {
-        receiveFirstMaterialCard();
-      }
-      view.display();
+void PlayView::printTownPlayer(list<Vertex*>::iterator vIterator, int x,
+    int y) {
+    if (game.graph.getVertex((*vIterator)->getVertexId())->getOwner() == NULL) {
+            if ((*game.playerIterator)->towns->size() > 1) {
+                if ((*game.playerIterator)->getTownFirstTurn() < 2) {
+                    initializeIteratorTownList();
+                        setIsClickedToVertexGraph(vIterator);
+                        if (townIterator != (*game.playerIterator)->towns->end())
+                            printImages((*townIterator)->getImagePath(),
+                                (*vIterator)->getTown()->getPosX(),
+                                (*vIterator)->getTown()->getPosY());
+                        setOwnerToVertexGraph(game.graph.getVertex((*vIterator)->getVertexId()));
+                        deleteTowntoPlayer();
+                        (*game.playerIterator)->setScore(1);
+                        (*game.playerIterator)->setTownFirstTurn(1);
+                    if ((*game.playerIterator)->getTownFirstTurn() == 2) {
+                        receiveFirstMaterialCard();
+                    }
+                    game.graph.getVertex((*vIterator)->getVertexId())->setIsCity(false);
+                    game.graph.getVertex((*vIterator)->getVertexId())->setIsTown(true);
+                    view.display();
+                }
+            }
+            (*game.playerIterator)->setFirstTurnFinished(true);
     }
-  }
-  (*game.playerIterator)->setFirstTurnFinished(true);
 }
 
 void PlayView::searhTown(double x, double y, list<Land*>::iterator it) {
@@ -403,22 +439,50 @@ void PlayView::searhTown(double x, double y, list<Land*>::iterator it) {
             if (selectTown) {
                 buyTown(vIterator, (int)x, (int)y);
             }
+            if (selectCity) {
+                buyCity(vIterator, (int)x, (int)y);
+            }
         }
         vIterator++;
+    }
+}
+void PlayView::buyCity(list<Vertex*>::iterator vIterator, double x,
+    double y) {
+    initializeIteratorCityList();
+    setIsClickedToVertexGraph(vIterator);
+    if (game.graph.getVertex((*vIterator)->getVertexId())->getOwner() == NULL) {
+        if ((*game.playerIterator)->towns->size() > 1) {///arreglar este BUG
+            if (townIterator != (*game.playerIterator)->towns->end())
+                printImages((*cityIterator)->getImagePath(),
+                    (*vIterator)->getTown()->getPosX(),
+                    (*vIterator)->getTown()->getPosY());
+            setOwnerToVertexGraph(game.graph.getVertex((*vIterator)->getVertexId()));
+            deleteCitytoPlayer();
+            (*game.playerIterator)->setScore(2);
+            view.display();
+            selectCity = false;
+            game.graph.getVertex((*vIterator)->getVertexId())->setIsCity(true);
+            game.graph.getVertex((*vIterator)->getVertexId())->setIsTown(false);
+        }
     }
 }
 void PlayView::buyTown(list<Vertex*>::iterator vIterator, double x, double y) {
     initializeIteratorTownList();
     setIsClickedToVertexGraph(vIterator);
-    if ((*game.playerIterator)->towns->size() > 1) {///arreglar este BUG
-        if (townIterator != (*game.playerIterator)->towns->end())
-            printImages((*townIterator)->getImagePath(),
-                (*vIterator)->getTown()->getPosX(),
-                (*vIterator)->getTown()->getPosY());
-        setOwnerToVertexGraph(game.graph.getVertex((*vIterator)->getVertexId()));
-        deleteTowntoPlayer();
-        view.display();
-        selectTown = false;
+    if (game.graph.getVertex((*vIterator)->getVertexId())->getOwner() == NULL) {
+        if ((*game.playerIterator)->towns->size() > 1) {///arreglar este BUG
+            if (townIterator != (*game.playerIterator)->towns->end())
+                printImages((*townIterator)->getImagePath(),
+                    (*vIterator)->getTown()->getPosX(),
+                    (*vIterator)->getTown()->getPosY());
+            setOwnerToVertexGraph(game.graph.getVertex((*vIterator)->getVertexId()));
+            deleteTowntoPlayer();
+            (*game.playerIterator)->setScore(1);
+            view.display();
+            selectTown = false;
+            game.graph.getVertex((*vIterator)->getVertexId())->setIsCity(false);
+            game.graph.getVertex((*vIterator)->getVertexId())->setIsTown(true);
+        }
     }
 }
 void PlayView::receiveFirstMaterialCard() {
@@ -460,7 +524,6 @@ void PlayView::loadCardsRectangle() {
   cardsRectangle = playerRectangle;
   cardsRectangle.setPosition(350, 600);
   cardsRectangle.setSize({600, 100});
-  
   view.draw(cardsRectangle);
 }
 void PlayView::createBuyRectangle() {
@@ -483,6 +546,12 @@ void PlayView::printBuyDevelopCard() {
     printImages("Images/playerCard/mini_progressCard1.png", 220, 535);
     printImages("Images/playerCard/mini_knightCard1.png", 230, 535);
     printImages("Images/playerCard/mini_victoryPointCard1.png", 240, 535);
+}
+void PlayView::clickInDevelopCardBuy(int x, int y) {
+    if (x > 220 && x < 240+30 && y>535 && y < 535+30) {
+        BuyView buyView = BuyView();
+        buyView.goView();
+    }
 }
 void PlayView::clickInTownBuy(int x, int y) {
     if (x > 165 && x < 165 + 30 && y>535 && y < 535 + 30) {
@@ -510,7 +579,14 @@ void PlayView::payRawMaterialsToBuyTown() {
      //deleteWheattoPlayer();
      selectTown = true;
 }
-
+void PlayView::payRawMaterialsToBuyCity() {
+    int iterator = 0;
+    for (iterator; iterator < 3; iterator++) {
+        deleteWheattoPlayer();
+        deleteMineraltoPlayer();
+    }
+    selectCity = true;
+}
 
 
 void PlayView::deleteClaytoPlayer() {
@@ -538,9 +614,7 @@ void PlayView::deleteWheattoPlayer() {
     if (wheatIterator != (*game.playerIterator)->wheatlCard->end())
         (*game.playerIterator)->wheatlCard->pop_back();
 }
-void PlayView::payRawMaterialsToBuyCity() {
 
-}
 void PlayView::buildTown() {
     if ((*game.playerIterator)->clayCard->size() > 1 && (*game.playerIterator)->woodCard->size() > 1) {
         //&& (*game.playerIterator)->wheatlCard->size() > 1 && (*game.playerIterator)->woolCard->size() > 1)
@@ -553,7 +627,7 @@ void PlayView::buildCity() {
         payRawMaterialsToBuyCity();
 }
 void PlayView::isBuyButtonClicked(sf::Event event) {
-    
+    //if (buy.isPressed(event)) {
         if (isTownBuyClicked) {
             buildTown();
         }
@@ -593,7 +667,9 @@ void PlayView::drawView() {
   printPlayerCard();
   drawLabelCardPlayer();
   drawLabelFigurePlayer();
+  drawLabelDevelopCardPLayer();
   printPlayerFigure();
+  printPlayerDevelopCard();
   createButtons();
   printPlayerBuyFigure();
   printBuyDevelopCard();
@@ -605,7 +681,9 @@ void PlayView::drawView() {
 void PlayView::initializeIteratorTownList() {
   townIterator = (*game.playerIterator)->towns->begin();
 }
-
+void PlayView::initializeIteratorCityList() {
+    cityIterator = (*game.playerIterator)->citys->begin();
+}
 void PlayView::setIsClickedToVertexGraph(list<Vertex *>::iterator it) {
   game.graph.getVertex((*it)->getVertexId())->setClicked(true);
 }
@@ -655,7 +733,10 @@ void PlayView::deleteTowntoPlayer() {
   if (townIterator != (*game.playerIterator)->towns->end())
     (*game.playerIterator)->towns->pop_back(); // poner alerta aca
 }
-
+void PlayView::deleteCitytoPlayer() {
+    if (cityIterator != (*game.playerIterator)->citys->end())
+        (*game.playerIterator)->citys->pop_back(); // poner alerta aca
+}
 void PlayView::setOwnerToVertexGraph(Vertex *vertex) {
   game.graph.getVertex(vertex->getVertexId())->setOwner((*game.playerIterator));
 }
@@ -704,18 +785,11 @@ void PlayView::goView() {
                   isDiceButtonClicked(getMousePositionX(view), getMousePositionY(view));
                   clickInCityBuy(getMousePositionX(view), getMousePositionY(view));
                   clickInTownBuy(getMousePositionX(view), getMousePositionY(view));
-                  if (buy.isPressed(event)) {
-                      isBuyButtonClicked(eventTest);
-                      break;
-                  }
+                  clickInDevelopCardBuy(getMousePositionX(view), getMousePositionY(view));
+                  isBuyButtonClicked(eventTest);
               }
               isTurnButtonClicked(sf::Mouse::getPosition(view).x,
                   sf::Mouse::getPosition(view).y);
-              if (buy.isMouseOver(view)) {
-
-                  BuyView buyView = BuyView();
-                  buyView.goView();
-            }
           }
           break;
         case sf::Event::Closed:
@@ -1136,14 +1210,26 @@ void PlayView::loadHexagonNodes(list<Vertex *>::iterator itX, double posX,
     }
     else if (!getIsVertexGraphPrinted(itX)) {
         game.graph.getVertex((*itX)->getVertexId())->setPrinted(true);
-        townIterator =
-            game.graph.getVertex((*itX)->getVertexId())
-            ->getOwner()
-            ->towns->begin(); //(*game.playerIterator)->towns->begin();
-        if (townIterator !=
-            game.graph.getVertex((*itX)->getVertexId())->getOwner()->towns->end()) {
-            printImages((*townIterator)->getImagePath(), relativePositionX,
-                relativePositionY);
+        if (game.graph.getVertex((*itX)->getVertexId())->getIsTown()) {
+            townIterator =
+                game.graph.getVertex((*itX)->getVertexId())
+                ->getOwner()
+                ->towns->begin(); //(*game.playerIterator)->towns->begin();
+            if (townIterator !=
+                game.graph.getVertex((*itX)->getVertexId())->getOwner()->towns->end()) {
+                printImages((*townIterator)->getImagePath(), relativePositionX,
+                    relativePositionY);
+            }
+        }
+        if (game.graph.getVertex((*itX)->getVertexId())->getIsCity()) {
+            cityIterator = game.graph.getVertex((*itX)->getVertexId())
+                ->getOwner()
+                ->citys->begin();
+            if (cityIterator !=
+                game.graph.getVertex((*itX)->getVertexId())->getOwner()->citys->end()) {
+                printImages((*cityIterator)->getImagePath(), relativePositionX,
+                    relativePositionY);
+            }
         }
     }
 }
